♻️ 集計機能を関数に切り出し、aggregation.tsに移動
