+++ packages/fukui-terminal/src/App.tsx
@@ -9,17 +9,15 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { AggregatedData } from "@/interfaces/aggregated-data.interface";
+import {
+  aggregateDaily,
+  aggregateHourly,
+  aggregateMonthly,
+  aggregateWeekly,
+} from "@/lib/aggregation";
 import { getDailyData, getData } from "@/lib/data/csv";
 import { useEffect, useState } from "react";
-import * as holidayJP from "@holiday-jp/holiday_jp";
 import { LoadingSpinner } from "./components/parts/loading-spinner";
-
-function filterByRange(data: AggregatedData[], from: Date, to: Date) {
-  return data.filter((row) => {
-    const date = new Date(row["aggregate from"]);
-    return date >= from && date <= to;
-  });
-}
 
 function App() {
   useEffect(() => {
@@ -121,124 +119,28 @@ function App() {
   }, [theme, startDate, endDate]);
 
   useEffect(() => {
-    let filtered = csvData;
+    const filtered = csvData;
     const filteredDaily = csvDailyData;
 
     if (theme === "month" && startMonth && endMonth) {
       // 月末を取得
       const end = new Date(endMonth.getFullYear(), endMonth.getMonth() + 1, 0);
-
-      filtered = filterByRange(filtered, startMonth, end);
-
-      // 月ごとに集計
-      const monthlyMap = new Map<string, AggregatedData>();
-      filtered.forEach((row) => {
-        const date = new Date(row["aggregate from"]);
-        const monthKey = `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, "0")}`;
-        if (!monthlyMap.has(monthKey)) {
-          monthlyMap.set(monthKey, {
-            ...row,
-            ["aggregate from"]: `${monthKey}`,
-            ["aggregate to"]: `${monthKey}`,
-            ["total count"]: Number(row["total count"]),
-          });
-        } else {
-          const prev = monthlyMap.get(monthKey)!;
-          monthlyMap.set(monthKey, {
-            ...prev,
-            ["total count"]: Number(prev["total count"]) + Number(row["total count"]),
-          });
-        }
-      });
-      setFilteredData(Array.from(monthlyMap.values()));
+      setFilteredData(aggregateMonthly(csvData, startMonth, end));
       return;
     }
 
     if (theme === "week" && startWeekRange && endWeekRange) {
-      filtered = filterByRange(filtered, startWeekRange.from, endWeekRange.to);
-
-      const weeklyAggregated: AggregatedData[] = [];
-      let i = 0;
-      while (i < filtered.length) {
-        let weekRows;
-        if (i === 0) {
-          // 最初の週はstartWeekRange.from〜startWeekRange.toまで
-          weekRows = filtered.filter((row) => {
-            const d = new Date(row["aggregate from"]);
-            return d >= startWeekRange.from && d <= startWeekRange.to;
-          });
-          i += weekRows.length;
-        } else {
-          // 以降は7日ごと
-          weekRows = filtered.slice(i, i + 7);
-          i += 7;
-        }
-        const formatDate = (date: Date) =>
-          `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, "0")}-${String(date.getDate()).padStart(2, "0")}`;
-        const total = weekRows.reduce((sum, row) => sum + Number(row["total count"]), 0);
-        weeklyAggregated.push({
-          ...weekRows[0],
-          ["aggregate from"]: `${formatDate(new Date(weekRows[0]["aggregate from"]))}〜`,
-          ["aggregate to"]: `${formatDate(new Date(weekRows[weekRows.length - 1]["aggregate from"]))}`,
-          ["total count"]: total,
-        });
-      }
-      setFilteredData(weeklyAggregated);
+      setFilteredData(aggregateWeekly(csvData, startWeekRange, endWeekRange));
       return;
     }
 
     if (theme === "day" && startDate && endDate) {
-      filtered = filterByRange(filtered, startDate, endDate);
-
-      // 日ごとに集計
-      const dailyMap = new Map<string, AggregatedData>();
-      filtered.forEach((row) => {
-        const date = new Date(row["aggregate from"]);
-        const dayKey = `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, "0")}-${String(date.getDate()).padStart(2, "0")}`;
-        if (!dailyMap.has(dayKey)) {
-          const dayOfWeek = ["日", "月", "火", "水", "木", "金", "土"][date.getDay()];
-          const holiday = holidayJP.isHoliday(date);
-          let holidayName: string = "";
-          if (holiday) {
-            // 祝日一覧から該当日を検索
-            holidayName = holidayJP.between(date, date)[0].name;
-          }
-          dailyMap.set(dayKey, {
-            ...row,
-            ["aggregate from"]: `${dayKey}`,
-            ["aggregate to"]: `${dayKey}`,
-            ["total count"]: Number(row["total count"]),
-            dayOfWeek,
-            holidayName,
-          });
-        }
-      });
-      setFilteredData(Array.from(dailyMap.values()));
+      setFilteredData(aggregateDaily(csvData, startDate, endDate));
       return;
     }
 
     if (theme === "hour" && startDate && endDate) {
-      // 時間ごとに集計
-      const hourlyMap = new Map<string, AggregatedData>();
-      filteredDaily.forEach((row) => {
-        const date = new Date(row["aggregate from"]);
-        const hourKey = `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, "0")}-${String(date.getDate()).padStart(2, "0")} ${String(date.getHours()).padStart(2, "0")}:00`;
-        if (!hourlyMap.has(hourKey)) {
-          hourlyMap.set(hourKey, {
-            ...row,
-            ["aggregate from"]: hourKey,
-            ["aggregate to"]: hourKey,
-            ["total count"]: Number(row["total count"]),
-          });
-        } else {
-          const prev = hourlyMap.get(hourKey)!;
-          hourlyMap.set(hourKey, {
-            ...prev,
-            ["total count"]: Number(prev["total count"]) + Number(row["total count"]),
-          });
-        }
-      });
-      setFilteredDailyData(Array.from(hourlyMap.values()));
+      setFilteredDailyData(aggregateHourly(filteredDaily));
       return;
     }
